# Speed up OptimizedScalarQuantizer