--- a/docs/changelog/131599.yaml
+++ b/docs/changelog/131599.yaml
@@ -1,0 +1,5 @@
+pr: 131599
+summary: Speed up `OptimizedScalarQuantizer`
+area: Vector Search
+type: enhancement
+issues: []