… benchmarks/src/main/java/org/elasticsearch/benchmark/vector/OptimizedScalarQuantizerBenchmark.java
… libs/simdvec/src/test/java/org/elasticsearch/simdvec/ESVectorUtilTests.java
… server/src/main/java/org/elasticsearch/index/codec/vectors/es818/ES818BinaryQuantizedVectorsWriter.java
… server/src/test/java/org/elasticsearch/index/codec/vectors/OptimizedScalarQuantizerTests.java